+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `ChangemakerFieldValueBatch` entity type now exists to group changemaker field values that were imported together from an external source.
+- `POST /changemakerFieldValueBatches` endpoint now exists to create changemaker field value batches.
+
 ## 0.26.0 2025-10-30
 
 ### Added